Don't re-link SSL stuff if it's already there

+++ install.sh
@@ -75,8 +75,12 @@ fi
 
 # TEMPORARY: Link /etc/ssl/cert.pem to $NIX_SSL_CERT_FILE (for curl)
 #   Issue: https://github.com/NixOS/nixpkgs/issues/8247
-log "TEMPORARY: linking /etc/ssl/cert.pem to ${NIX_SSL_CERT_FILE} - sorry, requires sudo"
-sudo -- sh -c "mkdir -p /etc/ssl; ln -s ${NIX_SSL_CERT_FILE} /etc/ssl/cert.pem"
+if [ -e /etc/ssl/cert.pem ]; then
+    log "/etc/ssl/cert.pem exists already; not touching it."
+else
+    log "TEMPORARY: linking /etc/ssl/cert.pem to ${NIX_SSL_CERT_FILE} - sorry, requires sudo"
+    sudo -- sh -c "mkdir -p /etc/ssl; ln -s ${NIX_SSL_CERT_FILE} /etc/ssl/cert.pem"
+fi
 
 # Install oh-my-zsh if necessary
 if [ ! -d "${HOME}/.oh-my-zsh" ]; then
